Move adaptive sampling code out of internal pkg
