+++ plugin/sampling/calculationstrategy/interface.go
