+++ plugin/sampling/leaderelection/mocks/ElectionParticipant.go
