+++ plugin/sampling/strategystore/adaptive/processor_test.go
@@ -30,8 +30,8 @@ import (
 	"github.com/jaegertracing/jaeger/cmd/collector/app/sampling/model"
 	jio "github.com/jaegertracing/jaeger/pkg/io"
 	"github.com/jaegertracing/jaeger/pkg/testutils"
-	"github.com/jaegertracing/jaeger/plugin/sampling/internal/calculationstrategy"
-	epmocks "github.com/jaegertracing/jaeger/plugin/sampling/internal/leaderelection/mocks"
+	"github.com/jaegertracing/jaeger/plugin/sampling/calculationstrategy"
+	epmocks "github.com/jaegertracing/jaeger/plugin/sampling/leaderelection/mocks"
 	smocks "github.com/jaegertracing/jaeger/storage/samplingstore/mocks"
 	"github.com/jaegertracing/jaeger/thrift-gen/sampling"
 )
